Start audio preprocessing refactor. Created audio_preprocessing.py and moved a few appropriate classes in.
