+++ collections/nemo_asr/nemo_asr/__init__.py
@@ -14,11 +14,11 @@
 # ==============================================================================
 from nemo.core import Backend
 
+from .audio_preprocessing import (MultiplyBatch, SpectrogramAugmentation)
+from .beam_search_decoder import BeamSearchDecoderWithLM
 from .data_layer import (AudioToTextDataLayer, AudioPreprocessing,
-                         SpectrogramAugmentation, MultiplyBatch,
                          KaldiFeatureDataLayer, TranscriptDataLayer)
 from .greedy_ctc_decoder import GreedyCTCDecoder
-from .beam_search_decoder import BeamSearchDecoderWithLM
 from .jasper import JasperEncoder, JasperDecoderForCTC
 from .las.misc import JasperRNNConnector
 from .losses import CTCLossNM
