+++ collections/nemo_asr/setup.py
@@ -24,6 +24,7 @@ setuptools.setup(
         'num2words',
         'inflect',
         'kaldi-io',
+        'torchaudio',
         'torch-stft',
         'soundfile',
         'marshmallow',
